iyilestirme calismasi yapildi.
